+++ packages/react-image/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @suspensive/react-image
 
+## 0.0.22
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @suspensive/utils@2.10.0
+
+## 0.0.21
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @suspensive/utils@2.9.4
+
 ## 0.0.20
 
 ### Patch Changes