+++ packages/react/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @suspensive/react
 
+## 2.10.0
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @suspensive/utils@2.10.0
+
+## 2.9.4
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @suspensive/utils@2.9.4
+
 ## 2.9.3
 
 ### Patch Changes